neutron/orphan_tool/requirements.txt
python-keystoneclient
python-neutronclient